feat: Add state to App component
In this commit, I added state called todo to the app component.
todos is an array of todo items/objects which is then used for creating individual todo.
The todos slice of state is passed as prop into <TodoList /> component.
Within the <TodoLis /> component, the todos array is then mapped over and used to create a unique todo item